Blockstack ID calculator
Here's a program I wrote that determines the expiration date of your Blockstack ID. I'm looking to incorporate the actual Blockstack login and then having that read the information on the profile to determine your ID's expiration date. I also want to have a "Green" checkbox appear in the form when the ID you're searching for has been registered, and a "red x" when the ID you're searching for hasn't been registered.